Fix for tput with uninformative output
The tput installed on winbuilder gives an 'unknown terminal "cygwin"'
error on STDERR, but returns nothing on STDOUT and returns an error code
of 0.  Therefore the result cannot be coerced to a number and
cols returns character(0).

`character(0) %in% c(0, 1)` returns `logical(0)`, and `logical(0)` in an if
statement yields the fairly useless error message 'error: argument is of
length zero'

This situation results in crayon being unable to load.  The fix handles
the above situation as well as the situation where tput does return
something on STDOUT, but that value cannot be coerced to a number (and
returns NA).

Fixes #18